+++ gempy/library/cython_utils.pyx
@@ -145,21 +145,22 @@ cdef void mask_stats(float data[], unsigned short mask[], int has_mask,
     result[1] = sumsq / nused - mean*mean
 
 
-cdef long num_good(unsigned short mask[], long data_size):
+@cython.exceptval(check=False)
+cdef long num_good(unsigned short mask[], int data_size) except? 0:
     """
     Returns the number of unmasked pixels in an array.
 
     Parameters
     ----------
     mask : unsigned short array
-        (add description)
+        array of mask pixels
     data_size : long
-        (add description)
+        size of array
 
     Returns
     -------
-    ngood : (?)
-        (add description)
+    ngood : long
+        number of good pixels in stack
     """
     cdef long i, ngood = 0
 
@@ -173,7 +174,7 @@ cdef long num_good(unsigned short mask[], long data_size):
 @cython.boundscheck(False)
 @cython.wraparound(False)
 def iterclip(float [:] data, unsigned short [:] mask, float [:] variance,
-             int has_var, int num_img, long data_size, double lsigma,
+             int has_var, long num_img, long data_size, double lsigma,
              double hsigma, int max_iters, int mclip, int sigclip):
     """
     Iterative sigma-clipping. This is the function that interfaces with python.
@@ -188,7 +189,7 @@ def iterclip(float [:] data, unsigned short [:] mask, float [:] variance,
         (add description)
     has_var : int
         Worry about the input variance array?
-    num_img : int
+    num_img : long
         Number of input images.
     data_size : long
         Number of pixels per input image
@@ -294,3 +295,113 @@ def landstat(double [:] landscape, int [:] coords, int [:] len_axes,
             sum += landscape[l]
 
     return sum
+
+##############################################################
+# The following code is used by polynomial inerpolators
+
+@cython.boundscheck(False)
+@cython.wraparound(False)
+def polyinterp(float [:] array_in, int [:] lengths, int ndim,
+               float [:] array_out, long npixout,
+               float [:] geomap, int order, float oob_val):
+    """
+    Perform polynomial interpolation of one array to another array of the
+    same dimensionality.
+
+    Parameters
+    ----------
+    array_in : float array
+        input array that is being transformed
+    lengths : int array
+        axis lengths (in python order)
+    ndim : int
+        number of dimensions of array
+    array_out : float array
+        pre-created output array for placing results
+    npixout : long
+        size of output array
+    geomap : float array
+        ndim * npixout array representing locations in array_in where
+        each pixel in array_out maps back to
+    order : int (3 or 5)
+        order of interpolation (will be downgraded near boundaries)
+    oob_val : float
+        value to insert for out-of-bounds pixels
+    """
+    cdef long i, nstep
+    cdef int j, k, dim, totpix, num_pix=1
+    cdef float p, q, pp, qq, ppp, qqq, coord, sum
+    cdef float tmpcoeff[6]
+    cdef int pix1, npix
+
+    for dim in range(ndim):
+        num_pix *= order + 1
+
+    # space for all coefficients
+    cdef float *coeffs = <float *> malloc(num_pix * sizeof(float))
+    if not coeffs:
+        raise MemoryError()
+
+    # space for pixel number indices
+    cdef long *pixels = <long *> malloc(num_pix * sizeof(long))
+    if not pixels:
+        raise MemoryError()
+
+    for i in range(npixout):
+        for j in range(num_pix):
+            coeffs[j] = 1.0
+            pixels[j] = 0
+
+        totpix = 1  # total number of input pixels contributing
+        nstep = 1  # indexing increment in this dimension
+        for dim in range(ndim-1, -1, -1):
+            if totpix > 0:
+                coord = geomap[dim * npixout + i]
+                if coord >= 0 and coord < lengths[dim] - 1:
+                    pix1 = int(coord)
+                    p = coord - pix1
+                    q = 1. - p
+                    if pix1 > 0 and pix1 < lengths[dim] - 2 and order >= 3:
+                        pp = p * (p * p - 1.) / 6.
+                        qq = q * (q * q - 1.) / 6.
+                        if pix1 > 1 and pix1 < lengths[dim] - 3 and order >= 5:
+                            ppp = pp * (p * p - 4.) / 20.
+                            qqq = qq * (q * q - 4.) / 20.
+                            pix1 -= 2
+                            npix = 6
+                            tmpcoeff[0] = qqq
+                            tmpcoeff[1] = qq - 4 * qqq + ppp
+                            tmpcoeff[2] = q - qq - qq + pp + 6 * qqq - 4 * ppp
+                            tmpcoeff[3] = p - pp - pp + qq + 6 * ppp - 4 * qqq
+                            tmpcoeff[4] = pp - 4 * ppp + qqq
+                            tmpcoeff[5] = ppp
+                        else:  # cubic interpolation
+                            pix1 -= 1
+                            npix = 4
+                            tmpcoeff[0] = qq
+                            tmpcoeff[1] = q + pp - qq - qq
+                            tmpcoeff[2] = p + qq - pp - pp
+                            tmpcoeff[3] = pp
+                    else:  # only linear interpolation
+                        npix = 2
+                        tmpcoeff[0] = q
+                        tmpcoeff[1] = p
+
+                    for j in range(totpix):
+                        for k in range(npix-1, -1, -1):
+                            pixels[k*totpix + j] = pixels[j] + (pix1 + k) * nstep
+                            coeffs[k*totpix + j] = coeffs[j] * tmpcoeff[k]
+
+                    totpix *= npix
+                    nstep *= lengths[dim]
+
+                else:
+                    totpix = 0
+
+        if totpix > 0:
+            sum = 0.
+            for j in range(totpix):
+                sum += coeffs[j] * array_in[pixels[j]]
+            array_out[i] = sum
+        else:
+            array_out[i] = oob_val